refactor!: rename dkg directory to mpc
