+++ crates/hashi/src/communication/sui_tob.rs
@@ -13,8 +13,8 @@ use super::ChannelError;
 use super::ChannelResult;
 use super::OrderedBroadcastChannel;
 use crate::config::HashiIds;
-use crate::dkg::types::CertificateV1;
-use crate::dkg::types::DealerMessagesHash;
+use crate::mpc::types::CertificateV1;
+use crate::mpc::types::DealerMessagesHash;
 use crate::onchain::OnchainState;
 use crate::sui_tx_executor::SuiTxExecutor;
 use hashi_types::committee::Committee;
