+++ crates/hashi/src/grpc/client.rs
@@ -5,14 +5,14 @@ use tonic::Response;
 use tonic_rustls::Channel;
 use tonic_rustls::Endpoint;
 
-use crate::dkg::types::ComplainRequest;
-use crate::dkg::types::ComplaintResponses;
-use crate::dkg::types::GetPublicDkgOutputRequest;
-use crate::dkg::types::GetPublicDkgOutputResponse;
-use crate::dkg::types::RetrieveMessagesRequest;
-use crate::dkg::types::RetrieveMessagesResponse;
-use crate::dkg::types::SendMessagesRequest;
-use crate::dkg::types::SendMessagesResponse;
+use crate::mpc::types::ComplainRequest;
+use crate::mpc::types::ComplaintResponses;
+use crate::mpc::types::GetPublicDkgOutputRequest;
+use crate::mpc::types::GetPublicDkgOutputResponse;
+use crate::mpc::types::RetrieveMessagesRequest;
+use crate::mpc::types::RetrieveMessagesResponse;
+use crate::mpc::types::SendMessagesRequest;
+use crate::mpc::types::SendMessagesResponse;
 use crate::tls::make_client_config_no_verification;
 use hashi_types::proto::GetReconfigCompletionSignatureRequest;
 use hashi_types::proto::GetServiceInfoRequest;
