+++ crates/hashi/src/mpc/mod.rs
@@ -1,0 +1,8 @@
+mod mpc_except_signing;
+pub mod rpc;
+pub mod service;
+pub mod types;
+
+pub use mpc_except_signing::*;
+pub use service::MpcHandle;
+pub use service::MpcService;
