+++ crates/hashi/src/mpc/service.rs
@@ -13,11 +13,11 @@ use tracing::warn;
 use crate::Hashi;
 use crate::communication::SuiTobChannel;
 use crate::communication::fetch_certificates;
-use crate::dkg::DkgManager;
-use crate::dkg::DkgOutput;
-use crate::dkg::rpc::RpcP2PChannel;
-use crate::dkg::types::CertificateV1;
-use crate::dkg::types::ProtocolType;
+use crate::mpc::DkgManager;
+use crate::mpc::DkgOutput;
+use crate::mpc::rpc::RpcP2PChannel;
+use crate::mpc::types::CertificateV1;
+use crate::mpc::types::ProtocolType;
 use crate::onchain::Notification;
 use crate::onchain::OnchainState;
 use fastcrypto_tbls::threshold_schnorr::G;
